requirements.txt
requests
pandas
python-dotenv
tqdm
streamlit
matplotlib
plotly
furl
